api/requirements.txt
click==8.1.7
Flask==2.2.5
Flask-Cors==3.0.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.17.0
